src/requirements.txt
bottle~=0.12.19
dateparser~=1.1.0
bs4~=0.0.1
beautifulsoup4~=4.10.0
requests~=2.26.0
psycopg2-binary~=2.9.1